image not update if empty send or undefined in banner and brand handeled

+++ src/config/helper.js
@@ -29,7 +29,7 @@ const getTokenFromHeader = (req) =>{
 }
 
 const deleteFile = (path,filename) =>{
-  fs.existsSync(path) && fs.unlinkSync(path+filename);
+  filename && fs.existsSync(path) && fs.unlinkSync(path+filename);
 }
